LayoutTests/platform/ios-sim-deprecated/media/video-currentTime-set-expected.txt

Test that setting currentTime changes the time, and that 'ended' event is fired in a reasonable amount of time

EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
RUN(video.currentTime = video.duration - 0.2)
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0'), OBSERVED '5.827200412750244' FAIL
RUN(video.currentTime = video.duration - 0.2)
EVENT(seeked)
EXPECTED (video.currentTime.toFixed(2) == '5.83'), OBSERVED '0.00' FAIL
RUN(video.play())

EVENT(ended)
END OF TEST
